refactor: 채팅 도메인 exception 및 code 패키지 구조 정리
