+++ src/main/java/com/capstone/pickIt/api/chat/service/ChatRoomCommandServiceImpl.java
@@ -1,10 +1,10 @@
 package com.capstone.pickIt.api.chat.service;
 
-import com.capstone.pickIt.api.chat.code.ChatErrorCode;
+import com.capstone.pickIt.domain.chat.exception.ChatErrorCode;
 import com.capstone.pickIt.api.chat.converter.ChatRoomConverter;
 import com.capstone.pickIt.api.chat.dto.request.DirectChatRoomCreateRequestDTO;
 import com.capstone.pickIt.api.chat.dto.response.DirectChatRoomResponseDTO;
-import com.capstone.pickIt.api.chat.exception.ChatException;
+import com.capstone.pickIt.domain.chat.exception.ChatException;
 import com.capstone.pickIt.domain.chat.entity.ChatPart;
 import com.capstone.pickIt.domain.chat.entity.ChatRoom;
 import com.capstone.pickIt.domain.chat.repository.ChatPartRepository;
