+++ src/main/java/com/capstone/pickIt/domain/chat/exception/ChatException.java
@@ -1,4 +1,4 @@
-package com.capstone.pickIt.api.chat.exception;
+package com.capstone.pickIt.domain.chat.exception;
 
 import com.capstone.pickIt.global.apiPayload.exception.CustomException;
 import com.capstone.pickIt.global.apiPayload.response.BaseCode;
@@ -6,9 +6,5 @@ import com.capstone.pickIt.global.apiPayload.response.BaseCode;
 public class ChatException extends CustomException {
     public ChatException(BaseCode errorCode) {
         super(errorCode);
-    }
-
-    public ChatException(BaseCode errorCode, String message) {
-        super(errorCode, message);
     }
 }
